Refresh Caches
There is now a user-interface (and a single-endpoint, unauthenticated API, for that matter) to refresh the cached data. One serendipitous aspect of this update is that the caching is now _more_ reliable, as a variety of updates to [battis\simplecache](https://github.com/battis/simplecache) were finalized to support this.